Reorganized code
Reorganized code, packages and subfolders for scalability
