+++ src/utils/utils.go
@@ -1,4 +1,4 @@
-package bp
+package utils
 
 import (
 	"os"
